Move NewDerivedKeyFromSeed to ext functions file and added sanity test
